# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger UI",
        "description": "L5 Swagger API description",
        "contact": {
            "name": "Dinh Huan",
            "email": "dinhhuan1357@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/products/{id}/variants": {
            "get": {
                "tags": [
                    "Variant"
                ],
                "summary": "Tìm mùi vị của sản phẩm",
                "operationId": "980bb81304dedea55e4352ca4888d2eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "mã sản phẩm",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tìm mùi vị thành công"
                    },
                    "400": {
                        "description": "Tìm mùi vị thất bại"
                    }
                }
            }
        },
        "/api/products/variants": {
            "post": {
                "tags": [
                    "Variant"
                ],
                "summary": "Tạo thêm mùi vị",
                "description": "Tạo mùi vị cho sản phẩm, yêu cầu có ảnh",
                "operationId": "9cba641a43dc51401220ad2910c28a09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "VariantName": {
                                        "description": "Tên mùi vị",
                                        "type": "string",
                                        "example": "Milk Tea Flavor"
                                    },
                                    "ProductID": {
                                        "description": "Mã sản phẩm",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "StockQuantity": {
                                        "description": "Số lượng sản phẩm",
                                        "format": "int32",
                                        "example": "40"
                                    },
                                    "Image": {
                                        "description": "File ảnh mùi vị",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo mùi vị thành công"
                    },
                    "400": {
                        "description": "Tạo mùi vị thất bại"
                    }
                }
            }
        },
        "/api/products/variants/{id}": {
            "delete": {
                "tags": [
                    "Variant"
                ],
                "summary": "xóa mùi vị",
                "description": "Xóa một mùi vị có {id}",
                "operationId": "a49dd4d50c0a3cdf6dd84fd36c0b08d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id của mùi  vị",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "xóa thành công"
                    },
                    "400": {
                        "description": "xóa thất bại"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Variant"
                ],
                "summary": "Cập nhật thông tin mùi vị",
                "description": "Cập nhật thông tin mùi vị có {id}",
                "operationId": "617bba168c269a952ee31826b1b2c317",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id mùi vị",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "VariantName",
                                    "ProductID",
                                    "StockQuantity"
                                ],
                                "properties": {
                                    "VariantName": {
                                        "type": "string",
                                        "example": "chocola"
                                    },
                                    "ProductID": {
                                        "type": "integer"
                                    },
                                    "StockQuantity": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "cập nhật thành công"
                    },
                    "400": {
                        "description": "cập nhật thất bại"
                    }
                }
            }
        },
        "/api/products/variants/image/{image_id}": {
            "patch": {
                "tags": [
                    "Variant"
                ],
                "summary": "cập nhập hình ảnh mùi vị",
                "description": "Cập nhật hình ảnh mùi vị có image_id",
                "operationId": "50f1438bb28ecfeded0a419c29eb48f2",
                "parameters": [
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "image id của mùi vị",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart-formdata": {
                            "schema": {
                                "properties": {
                                    "Image": {
                                        "description": "Ảnh thay thế",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật ảnh thành công"
                    },
                    "400": {
                        "description": "Cập nhật ảnh thất bại"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}